refactor scripts logic to get rid of lag
